Fix CI when the polyfill is inert.
Currently, CI is failing for beta and canary scenarios because we are
eagerly importing from `ember-on-modifier/modifiers/on` in
`tests/test-helper.js`, but when ran against Ember >= 3.11 we emit no
modules causing the import to fail.

This makes the import use `require` so that we can detect the Ember >=
3.11 case, and avoid erroring. The downside is that we can no longer
make count based assertions, but those assertions already live upstream
in Ember so it's not an issue.